Allow more subtle side sculpting of SA/DSA to make the tops less square.
To enable this, you need to set

$enable_more_side_sculpting = true

(as well as ensuring that the existing enable_side_sculpting is set).

There is also an additional method for doing the extra sculpting that
can be enabled via:

  $enable_more_side_sculpting = "slow"

This was actually my first attempt which looks nice at 1u but doesn't
scale nicely to wider/taller keys.